# api/swagger.yaml
...
swagger: "2.0"
info:
  description: "## TMF API Reference: TMF656 - Service Problem Management\n\n### Release\
    \ : 18.5 - December 2018\n\nService Problem Management API goal is to provide\
    \ the ability to manage problems in the Service domain. \n\n### Resource \n- serviceProblem\n\
    \n### Operations\nService Problem Management API performs the following operations\
    \ on the resource :\n- Retrieve an entity or a collection of entities depending\
    \ on filter criteria\n- Partial update of an entity (including updating rules)\n\
    - Create an entity (including default values and creation rules)\n- Delete an\
    \ entity (for administration purposes)\n- Manage notification of events"
  version: "4.0.0"
  title: "Service Problem Management"
host: "localhost:8080"
basePath: "/tmf-api"
...
schemes:
- "http"
consumes:
- "application/json;charset=utf-8"
produces:
- "application/json;charset=utf-8"
paths:
...
  /listener/serviceProblemCreateNotification:
    post:
      tags:
      - "notification listeners (client side)"
      summary: "Client listener for entity ServiceProblemCreateNotification"
      description: "Example of a client listener for receiving the notification ServiceProblemCreateNotification"
      operationId: "listenToServiceProblemCreateNotification"
      parameters:
      - in: "body"
        name: "data"
        description: "The event data"
        required: true
        schema:
          $ref: "#/definitions/ServiceProblemCreateNotification"
      responses:
        201:
          description: "Notified"
          schema:
            $ref: "#/definitions/EventSubscription"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Method Not allowed"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "NotificationListenersClientSide"
  /listener/serviceProblemStateChangeNotification:
    post:
      tags:
      - "notification listeners (client side)"
      summary: "Client listener for entity ServiceProblemStateChangeNotification"
      description: "Example of a client listener for receiving the notification ServiceProblemStateChangeNotification"
      operationId: "listenToServiceProblemStateChangeNotification"
      parameters:
      - in: "body"
        name: "data"
        description: "The event data"
        required: true
        schema:
          $ref: "#/definitions/ServiceProblemStateChangeNotification"
      responses:
        201:
          description: "Notified"
          schema:
            $ref: "#/definitions/EventSubscription"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Method Not allowed"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "NotificationListenersClientSide"
  /listener/serviceProblemAttributeValueChangeNotification:
    post:
      tags:
      - "notification listeners (client side)"
      summary: "Client listener for entity ServiceProblemAttributeValueChangeNotification"
      description: "Example of a client listener for receiving the notification ServiceProblemAttributeValueChangeNotification"
      operationId: "listenToServiceProblemAttributeValueChangeNotification"
      parameters:
      - in: "body"
        name: "data"
        description: "The event data"
        required: true
        schema:
          $ref: "#/definitions/ServiceProblemAttributeValueChangeNotification"
      responses:
        201:
          description: "Notified"
          schema:
            $ref: "#/definitions/EventSubscription"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Method Not allowed"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "NotificationListenersClientSide"
  /listener/serviceProblemInformationRequiredNotification:
    post:
      tags:
      - "notification listeners (client side)"
      summary: "Client listener for entity ServiceProblemInformationRequiredNotification"
      description: "Example of a client listener for receiving the notification ServiceProblemInformationRequiredNotification"
      operationId: "listenToServiceProblemInformationRequiredNotification"
      parameters:
      - in: "body"
        name: "data"
        description: "The event data"
        required: true
        schema:
          $ref: "#/definitions/ServiceProblemInformationRequiredNotification"
      responses:
        201:
          description: "Notified"
          schema:
            $ref: "#/definitions/EventSubscription"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Method Not allowed"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "NotificationListenersClientSide"
definitions:
...
  EventSubscription:
    type: "object"
    required:
    - "callback"
    - "id"
    properties:
      id:
        type: "string"
        description: "Id of the listener"
      callback:
        type: "string"
        description: "The callback being registered."
      query:
        type: "string"
        description: "additional data to be passed"
    description: "Sets the communication endpoint address the service instance must\
      \ use to deliver notification information"
    example:
      query: "query"
      callback: "callback"
      id: "id"
...
  ServiceProblemCreateNotification:
    type: "object"
    properties:
      eventId:
        type: "string"
        description: "The identifier of the notification"
      eventTime:
        type: "string"
        format: "date-time"
        description: "Time of the event occurence"
      eventType:
        type: "string"
        description: "The type of the notification"
      fieldPath:
        type: "string"
        description: "The path identifying the object field concerned by this notification"
      resourcePath:
        type: "string"
        description: "The path identifying the resource object concerned by this notification"
      event:
        description: "The event linked to the involved resource object"
        $ref: "#/definitions/ServiceProblemCreateEvent"
    description: "The notification data structure"
    example:
      eventId: "eventId"
      resourcePath: "resourcePath"
      eventTime: "2000-01-23T04:56:07.000+00:00"
      fieldPath: "fieldPath"
      eventType: "eventType"
      event:
        serviceProblem:
          statusChangeReason: "statusChangeReason"
          reason: "reason"
          trackingRecord:
          - systemId: "systemId"
            extensionInfo:
            - '@baseType': "@baseType"
              '@type': "@type"
              valueType: "valueType"
              name: "name"
              '@schemaLocation': "http://example.com/aeiou"
              value: {}
            - '@baseType': "@baseType"
              '@type': "@type"
              valueType: "valueType"
              name: "name"
              '@schemaLocation': "http://example.com/aeiou"
              value: {}
            '@baseType': "@baseType"
            '@type': "@type"
            description: "description"
            id: "id"
            time: "2000-01-23T04:56:07.000+00:00"
            '@schemaLocation': "http://example.com/aeiou"
            user: "user"
          - systemId: "systemId"
            extensionInfo:
            - '@baseType': "@baseType"
              '@type': "@type"
              valueType: "valueType"
              name: "name"
              '@schemaLocation': "http://example.com/aeiou"
              value: {}
            - '@baseType': "@baseType"
              '@type': "@type"
              valueType: "valueType"
              name: "name"
              '@schemaLocation': "http://example.com/aeiou"
              value: {}
            '@baseType': "@baseType"
            '@type': "@type"
            description: "description"
            id: "id"
            time: "2000-01-23T04:56:07.000+00:00"
            '@schemaLocation': "http://example.com/aeiou"
            user: "user"
          '@type': "@type"
          resolutionDate: "2000-01-23T04:56:07.000+00:00"
          responsibleParty:
            '@referredType': "@referredType"
            role: "role"
            '@baseType': "@baseType"
            '@type': "@type"
            name: "name"
            id: "id"
            href: "href"
            '@schemaLocation': "@schemaLocation"
          description: "description"
          underlyingProblem:
          - '@referredType': "@referredType"
            '@baseType': "@baseType"
            '@type': "@type"
            correlationId: "correlationId"
            id: "id"
            href: "href"
            '@schemaLocation': "http://example.com/aeiou"
          - '@referredType': "@referredType"
            '@baseType': "@baseType"
            '@type': "@type"
            correlationId: "correlationId"
            id: "id"
            href: "href"
            '@schemaLocation': "http://example.com/aeiou"
          statusChangeDate: "2000-01-23T04:56:07.000+00:00"
          problemEscalation: "problemEscalation"
          '@baseType': "@baseType"
          affectedService:
          - '@referredType': "@referredType"
            '@baseType': "@baseType"
            '@type': "@type"
            id: "id"
            href: "href"
            '@schemaLocation': "http://example.com/aeiou"
          - '@referredType': "@referredType"
            '@baseType': "@baseType"
            '@type': "@type"
            id: "id"
            href: "href"
            '@schemaLocation': "http://example.com/aeiou"
          associatedSLAViolation:
          - '@referredType': "@referredType"
            '@baseType': "@baseType"
            '@type': "@type"
            id: "id"
            href: "href"
            '@schemaLocation': "http://example.com/aeiou"
          - '@referredType': "@referredType"
            '@baseType': "@baseType"
            '@type': "@type"
            id: "id"
            href: "href"
            '@schemaLocation': "http://example.com/aeiou"
          timeChanged: "2000-01-23T04:56:07.000+00:00"
          correlationId: "correlationId"
          id: "id"
          href: "href"
          '@schemaLocation': "http://example.com/aeiou"
          impactPatterns:
            extensionInfo:
            - '@baseType': "@baseType"
              '@type': "@type"
              valueType: "valueType"
              name: "name"
              '@schemaLocation': "http://example.com/aeiou"
              value: {}
            - '@baseType': "@baseType"
              '@type': "@type"
              valueType: "valueType"
              name: "name"
              '@schemaLocation': "http://example.com/aeiou"
              value: {}
            '@baseType': "@baseType"
            '@type': "@type"
            description: "description"
            '@schemaLocation': "http://example.com/aeiou"
          timeRaised: "2000-01-23T04:56:07.000+00:00"
          underlyingAlarm:
          - '@referredType': "@referredType"
            '@baseType': "@baseType"
            '@type': "@type"
            id: "id"
            href: "href"
            changeRequest:
              '@referredType': "@referredType"
              '@baseType': "@baseType"
              '@type': "@type"
              id: "id"
              href: "href"
              '@schemaLocation': "http://example.com/aeiou"
            '@schemaLocation': "http://example.com/aeiou"
          - '@referredType': "@referredType"
            '@baseType': "@baseType"
            '@type': "@type"
            id: "id"
            href: "href"
            changeRequest:
              '@referredType': "@referredType"
              '@baseType': "@baseType"
              '@type': "@type"
              id: "id"
              href: "href"
              '@schemaLocation': "http://example.com/aeiou"
            '@schemaLocation': "http://example.com/aeiou"
          extensionInfo:
          - '@baseType': "@baseType"
            '@type': "@type"
            valueType: "valueType"
            name: "name"
            '@schemaLocation': "http://example.com/aeiou"
            value: {}
          - '@baseType': "@baseType"
            '@type': "@type"
            valueType: "valueType"
            name: "name"
            '@schemaLocation': "http://example.com/aeiou"
            value: {}
          relatedObject:
          - '@referredType': "@referredType"
            '@baseType': "@baseType"
            '@type': "@type"
            name: "name"
            id: "id"
            href: "href"
            '@schemaLocation': "http://example.com/aeiou"
          - '@referredType': "@referredType"
            '@baseType': "@baseType"
            '@type': "@type"
            name: "name"
            id: "id"
            href: "href"
            '@schemaLocation': "http://example.com/aeiou"
          rootCauseService:
          - '@referredType': "@referredType"
            '@baseType': "@baseType"
            '@type': "@type"
            id: "id"
            href: "href"
            '@schemaLocation': "http://example.com/aeiou"
          - '@referredType': "@referredType"
            '@baseType': "@baseType"
            '@type': "@type"
            id: "id"
            href: "href"
            '@schemaLocation': "http://example.com/aeiou"
          relatedEvent:
          - '@referredType': "@referredType"
            '@baseType': "@baseType"
            '@type': "@type"
            eventTime: "2000-01-23T04:56:07.000+00:00"
            id: "id"
            href: "href"
            eventType: "eventType"
            '@schemaLocation': "http://example.com/aeiou"
          - '@referredType': "@referredType"
            '@baseType': "@baseType"
            '@type': "@type"
            eventTime: "2000-01-23T04:56:07.000+00:00"
            id: "id"
            href: "href"
            eventType: "eventType"
            '@schemaLocation': "http://example.com/aeiou"
          originatorParty:
            '@referredType': "@referredType"
            role: "role"
            '@baseType': "@baseType"
            '@type': "@type"
            name: "name"
            id: "id"
            href: "href"
            '@schemaLocation': "@schemaLocation"
          priority: 6
          relatedParty:
          - '@referredType': "@referredType"
            role: "role"
            '@baseType': "@baseType"
            '@type': "@type"
            name: "name"
            id: "id"
            href: "href"
            '@schemaLocation': "@schemaLocation"
          - '@referredType': "@referredType"
            role: "role"
            '@baseType': "@baseType"
            '@type': "@type"
            name: "name"
            id: "id"
            href: "href"
            '@schemaLocation': "@schemaLocation"
          affectedResource:
          - '@referredType': "@referredType"
            '@baseType': "@baseType"
            '@type': "@type"
            name: "name"
            id: "id"
            href: "href"
            '@schemaLocation': "http://example.com/aeiou"
          - '@referredType': "@referredType"
            '@baseType': "@baseType"
            '@type': "@type"
            name: "name"
            id: "id"
            href: "href"
            '@schemaLocation': "http://example.com/aeiou"
          firstAlert:
            '@referredType': "@referredType"
            '@baseType': "@baseType"
            '@type': "@type"
            name: "name"
            id: "id"
            href: "href"
            '@schemaLocation': "http://example.com/aeiou"
          originatingSystem: "originatingSystem"
          associatedTroubleTicket:
          - '@referredType': "@referredType"
            '@baseType': "@baseType"
            '@type': "@type"
            correlationId: "correlationId"
            id: "id"
            href: "href"
            '@schemaLocation': "http://example.com/aeiou"
            status: "status"
          - '@referredType': "@referredType"
            '@baseType': "@baseType"
            '@type': "@type"
            correlationId: "correlationId"
            id: "id"
            href: "href"
            '@schemaLocation': "http://example.com/aeiou"
            status: "status"
          affectedNumberOfServices: 0
          impactImportanceFactor: "impactImportanceFactor"
          comment:
          - date: "2000-01-23T04:56:07.000+00:00"
            system: "system"
            '@baseType': "@baseType"
            author: "author"
            '@type': "@type"
            text: "text"
            '@schemaLocation': "http://example.com/aeiou"
          - date: "2000-01-23T04:56:07.000+00:00"
            system: "system"
            '@baseType': "@baseType"
            author: "author"
            '@type': "@type"
            text: "text"
            '@schemaLocation': "http://example.com/aeiou"
          category: "category"
          parentProblem:
          - '@referredType': "@referredType"
            '@baseType': "@baseType"
            '@type': "@type"
            correlationId: "correlationId"
            id: "id"
            href: "href"
            '@schemaLocation': "http://example.com/aeiou"
          - '@referredType': "@referredType"
            '@baseType': "@baseType"
            '@type': "@type"
            correlationId: "correlationId"
            id: "id"
            href: "href"
            '@schemaLocation': "http://example.com/aeiou"
          affectedLocation:
          - role: "role"
            '@baseType': "@baseType"
            '@type': "@type"
            name: "name"
            id: "id"
            href: "href"
            '@schemaLocation': "http://example.com/aeiou"
          - role: "role"
            '@baseType': "@baseType"
            '@type': "@type"
            name: "name"
            id: "id"
            href: "href"
            '@schemaLocation': "http://example.com/aeiou"
          status: "status"
          rootCauseResource:
          - '@referredType': "@referredType"
            '@baseType': "@baseType"
            '@type': "@type"
            name: "name"
            id: "id"
            href: "href"
            '@schemaLocation': "http://example.com/aeiou"
          - '@referredType': "@referredType"
            '@baseType': "@baseType"
            '@type': "@type"
            name: "name"
            id: "id"
            href: "href"
            '@schemaLocation': "http://example.com/aeiou"
...
  Error:
    type: "object"
    required:
    - "code"
    - "reason"
    properties:
      code:
        type: "integer"
        description: "Application relevant detail, defined in the API or a common\
          \ list."
      reason:
        type: "string"
        description: "Explanation of the reason for the error which can be shown to\
          \ a client user."
      message:
        type: "string"
        description: "More details and corrective actions related to the error which\
          \ can be shown to a client user."
      status:
        type: "integer"
        description: "HTTP Error code extension"
      referenceError:
        type: "string"
        format: "uri"
        description: "URI of documentation describing the error."
      '@baseType':
        type: "string"
        description: "When sub-classing, this defines the super-class."
      '@schemaLocation':
        type: "string"
        format: "uri"
        description: "A URI to a JSON-Schema file that defines additional attributes\
          \ and relationships"
      '@type':
        type: "string"
        description: "When sub-classing, this defines the sub-class entity name."
    description: "Used when an API throws an Error, typically with a HTTP error response-code\
      \ (3xx, 4xx, 5xx)"